requirements.txt
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.51.0
google-api-core==2.17.1
google-auth==2.28.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.56.4
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pandas==1.4.3
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
Pygments==2.17.2
pyodbc==4.0.39
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==25.1.2
regex==2022.8.17
requests==2.31.0
requests-oauthlib==1.4.0
requests-toolbelt==1.0.0
rsa==4.9
schedule==1.1.0
six==1.16.0
smartsheet-python-sdk==3.0.4
SQLAlchemy==1.4.25
sqlparse==0.4.4
stack-data==0.6.3
StrEnum==0.4.15
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.18.1
